requirements.txt
dnspython==2.2.1
checkdmarc